+++ resources/views/components/admin/blocks/rich-text.blade.php
@@ -2,6 +2,7 @@
 
 @php
     $c = "blocks.{$index}.content";
+    $b = '$wire.blocks['.json_encode((string) $index).'].content';
 @endphp
 
 <div class="flex flex-col gap-6">
@@ -21,13 +22,19 @@
     <div class="grid gap-4 md:grid-cols-2">
         <flux:radio.group wire:model.lazy="{{ $c }}.width" variant="segmented" label="{{ __('Width') }}">
             <flux:radio value="normal" label="{{ __('Normal') }}" />
-            <flux:radio value="narrow-left" label="{{ __('Narrow (left)') }}" />
-            <flux:radio value="narrow" label="{{ __('Narrow (centred)') }}" />
+            <flux:radio value="narrow" label="{{ __('Narrow') }}" />
         </flux:radio.group>
 
-        <flux:radio.group wire:model.lazy="{{ $c }}.align" variant="segmented" label="{{ __('Alignment') }}">
+        <flux:radio.group
+            wire:model.lazy="{{ $c }}.align"
+            variant="segmented"
+            label="{{ __('Alignment') }}"
+            x-show="{{ $b }}?.width === 'narrow'"
+            x-cloak
+        >
             <flux:radio value="left" icon="bars-3-bottom-left" label="{{ __('Left') }}" />
             <flux:radio value="center" icon="bars-3" label="{{ __('Center') }}" />
+            <flux:radio value="right" icon="bars-3-bottom-right" label="{{ __('Right') }}" />
         </flux:radio.group>
     </div>